🚚Rename: 프로젝트 구조 단순화하여 src에 파일 정리
